+++ src/Debug/Breakpoint/GhcFacade.hs
@@ -24,7 +24,43 @@ module Debug.Breakpoint.GhcFacade
   , pattern OverLit'
   ) where
 
-#if MIN_VERSION_ghc(9,4,0)
+#if MIN_VERSION_ghc(9,6,0)
+import           GHC.Driver.Plugins as Ghc hiding (TcPlugin)
+import           GHC.Hs.Extension as Ghc
+import           Language.Haskell.Syntax as Ghc
+import           GHC.Tc.Types as Ghc hiding (DefaultingPlugin)
+import qualified GHC.Tc.Plugin as Plugin
+import           GHC.Parser.Annotation as Ghc
+import           GHC.Types.SrcLoc as Ghc
+import           GHC.Types.Name as Ghc
+import           GHC.Iface.Env as Ghc
+import           GHC.Unit.Finder as Ghc
+import           GHC.Tc.Utils.Monad as Ghc hiding (TcPlugin, DefaultingPlugin)
+import           GHC.Data.FastString as Ghc
+import           GHC.Hs.Utils as Ghc
+import           GHC.Types.Unique.Set as Ghc
+import           GHC.Utils.Outputable as Ghc
+import           GHC.Hs.Binds as Ghc
+import           GHC.Rename.Bind as Ghc
+import           GHC.Data.Bag as Ghc
+import           GHC.Types.Basic as Ghc
+import           GHC.Types.Name.Env as Ghc
+import           GHC.Builtin.Types as Ghc
+import           GHC.Core.TyCo.Rep as Ghc
+import           GHC.Tc.Types.Constraint as Ghc
+import           GHC.Core.Make as Ghc
+import           GHC.Tc.Types.Evidence as Ghc
+import           GHC.Types.Id as Ghc
+import           GHC.Core.InstEnv as Ghc
+import           GHC.Core.Class as Ghc hiding (FunDep)
+import           GHC.Tc.Utils.TcType as Ghc
+import           GHC.Core.Type as Ghc
+import           GHC.Core.TyCon as Ghc
+import           GHC.Types.TyThing.Ppr as Ghc
+import           GHC.Hs.Expr as Ghc
+import           GHC.Types.PkgQual as Ghc
+
+#elif MIN_VERSION_ghc(9,4,0)
 import           GHC.Driver.Plugins as Ghc hiding (TcPlugin)
 import           GHC.Hs.Extension as Ghc
 import           Language.Haskell.Syntax as Ghc